dspbuffer: prevent unexpected disappearing of dsp element
While it is (in typical cases) unlikely that dsp element will change state
(to NULL) or even disappear before all output buffers have been finalized,
it is not impossible.  Consequently, dsp element should be kept around
until all buffers have "returned home", but upon their return the state
of the element may no longer be fully up and running, and has to be checked
for as well.

Signed-off-by: Mark Nauwelaerts <mark.nauwelaerts@collabora.co.uk>