dspbase: accept 0-sized buffer on padchain
If buffers with size 0 are passed to the dsp, it will return an error.
Those buffers are used in gstreamer 0.10 to force negotiation of elements
before the real flow starts. Often this is used to reduce flow startup latency.

dspbase now unrefs the buffer and returns success for those buffers.

Signed-off-by: Felipe Contreras <felipe.contreras@nokia.com>