base: simplify event list handling
The original commit b065dccc[1] seems to have introduced some issues.
See NB#286148.

One possibility suggested by Mark was to add a check to see if we are
flushing in sink_event(), and don't store the event in the previous
position.

However, there's no need to keep track of the previous position, or
store/retrieve events from multiple positions. Since now it's an
unlimited linked list, all events can be stored on the same position.

Storing events in multiple positions was just a hack anyway, and now
thanks to b065dccc, it's int needed any more.

This patch implements that, and ideally should have been part of
b065dccc.

[1] base: seralize all serialized events to output_loop

Signed-off-by: Felipe Contreras <felipe.contreras@nokia.com>